pyproject.toml
[project]
name = "pn-transcriptions"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "assemblyai>=0.46.0",
    "humanize>=4.10.0",
    "pytest>=9.0.1",
    "python-dotenv>=1.0.1",
    "rich>=13.9.2",
    "typer>=0.12.5",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.workspace]
members = ["scripts"]
